+++ tests/filter/filter.test.jsx
@@ -1,0 +1,108 @@
+import React, { PropTypes } from 'react';
+import chai, { expect } from 'chai';
+import chaiEnzyme from 'chai-enzyme';
+import { mount } from 'enzyme';
+
+/* Enzyme Helpers that can mount and unmount React component instances to
+ * the DOM and set `this.wrapper` and `this.dom` within Mocha's `this`
+ * context [full source here](tests/enzyme-helpers.js). `this` can
+ * only be referenced if inside `function () {}`.
+ */
+import { createMountNode, destroyMountNode } from '../enzyme-helpers';
+
+// Import your internal dependencies (for example):
+import Popover from '../../components/popover';
+import Button from '../../components/button';
+import Filter from '../../components/filter';
+
+/* Set Chai to use chaiEnzyme for enzyme compatible assertions:
+ * https://github.com/producthunt/chai-enzyme
+ */
+chai.use(chaiEnzyme());
+
+const defaultProps = {
+	id: 'sample-popover',
+	body: <span id="sample-body">This is the body</span>,
+	heading: <span id="sample-heading">This is the heading</span>
+};
+
+const defaultIds = {
+	trigger: defaultProps.id,
+	popover: `${defaultProps.id}-popover`,
+	body: `${defaultProps.id}-dialog-body`,
+	heading: `${defaultProps.id}-dialog-heading`
+};
+
+/* A re-usable demo component fixture outside of `describe` sections
+ * can accept props within each test and be unmounted after each tests.
+ * This wrapping component will be similar to your wrapping component
+ * you will create in the React Storybook for manual testing.
+ */
+const DemoComponent = React.createClass({
+	displayName: 'PopoverDemoComponent',
+	propTypes: {
+		isOpen: PropTypes.bool
+	},
+
+	getDefaultProps () {
+		return defaultProps;
+	},
+
+	getInitialState () {
+		return {};
+	},
+
+	render () {
+		return (
+			<Popover {...this.props}>
+				<Button label="Trigger Popover" />
+			</Popover>
+		);
+	}
+});
+
+/* All tests for component being tested should be wrapped in a root `describe`,
+ * which should be named after the component being tested.
+ * When read aloud, the cumulative `describe` and `it` names should form a coherent
+ * sentence, eg "Date Picker default structure and css is present with expected
+ * attributes set". If you are having trouble constructing a cumulative short
+ * sentence, this may be an indicator that your test is poorly structured.
+ * String provided as first parameter names the `describe` section. Limit to nouns
+ * as much as possible/appropriate.`
+ */
+describe('SLDSFilter', function () {
+	let mountNode;
+	let portalWrapper;
+	let wrapper;
+
+	// BASIC STRUCTURE
+
+	describe('Add custom props to Filter Popover', function () {
+		beforeEach(() => {
+			mountNode = createMountNode({ context: this });
+		});
+
+		afterEach(() => {
+			destroyMountNode({ wrapper, mountNode });
+		});
+
+		it('Filter could take popover as a prop and use the props of popover to render, verifies the custom popover className', (done) => {
+			const demoPopover = (<DemoComponent
+				className="custom-filter-popover"
+				isOpen
+				portalMount={(reactElement, domContainerNode) => {
+					portalWrapper = mount(reactElement, { attachTo: domContainerNode });
+				}}
+				onOpen={() => {
+					expect(portalWrapper.find('.custom-filter-popover')).to.exist;
+					done();
+				}}
+			/>);
+			wrapper = mount(<Filter
+				property="Show Me"
+				predicate="All Opportunities"
+				popover={demoPopover}
+			/>, { attachTo: mountNode });
+		});
+	});
+});